submissions/cezarfuhr/pix/pyproject.toml
[tool.poetry]
name = "conty-pix-challenge"
version = "0.1.0"
description = "API para o desafio de pagamentos em lote da Conty."
authors = ["Cezar Fuhr <cezar.fuhr@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pydantic-settings = "^2.10.1"
sqlalchemy = "^2.0.43"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
